refactor: Display meal queue in RecyclerView instead of TextView
Removed refresh button that was taking up screen space beneath the RecyclerView. Replaced the refresh button with a refresh ActionView in the options menu. This causes a indeterminate ProgressBar to become visible. Its visibility is hidden when we get a response from the server. Moving the control/visual-indicator to the options menu (action bar) allows for more screen space for the RecyclerView and the user is informed of the on-going network request without obscuring the main part of the screen.