feat: Define click listener for the meal queue
RecordOfMealAdapter's constructor now requires an implementation of OnItemClickListener, which is used to remove the item from the meal queue and notify the adapter that the data has changed.

A background is now defined for the RecyclerView item's layout (it changes the background color of the selected item). Also, added an ItemDecoration to the RecyclerView (inside MealQueueViewerActivity.onCreate(Bundle)), now there are horizontal bars between each item displayed in the RecyclerView.