feat: Track RV's currently selected index
Getting ready to apply LineTheCupWithCaramelCustomization using the selectedIndex that's being kept track of. A menu item will be selected (via clicking the RecyclerView's elements), and the new customization button have to distinguish between a compatible menu item or something that shouldn't be allowed to LineTheCupWithCaramelCustomization (e.g. a subclass of Food).