refactor: Display [MenuItem] using nested RecyclerView instead of TextView
Reverted [CustomizationDecorator] from decorator pattern where it was composed with a Drink object, to a POJO that all [MenuItem] can store as a List.

Click handling for the parent RecyclerView (the one for [MenuItem]) will update a [selectedIndex] field. Click handling for the child RecyclerView (the one for [CustomizationDecorator]) will remove the clicked item.