refactor: Update MealQueueViewerActivity to use nested RecyclerView instead of TextView
Also, updated converting a Meal to JSON and vice-versa. MealStagingActivity is able to send a Meal to the Kafka topic and MealQueueViewerActivity is able to pull from the Kafka topic and convert that response back into a Meal.

Using nested RecyclerView instead of TextView is allowing me to display the list of MenuItem (and their CustomizationDecorator if they have any), while enabling me to remove individual instances of MenuItem/CustomizationDecorator via clicking their respective View.