feat: Utilize RecyclerView.Adapter's [viewType] field in MenuItemAdapter
Reworked customizing a drink. Currently transitioning between all MenuItem having a list of CustomizationDecorator to all Drink having a list of Customization (each Customization subclass has an inner-class Builder and can save/load as JSONObject).

Have not applied the new Customization system to the view layer (learned to work with multiple ViewType within one RecyclerView). Currently, both ViewType options from MenuItemAdapter are exactly the same, except they set the background resource of the "name" TextView to different colors.

NEXT: Change one of the ViewType to display a list of Customization if the MenuItem is a Drink.