fix: Limit MenuItemInputFragment's child fragment manager to 2 fragment levels
The base level will be ViewPager2Fragment, and level above that can be one of the VerticalTextView's associated fragments (e.g. SyrupInputFragment, MilkInputFragment, and CustomizationInputFragment).

Previously, (1) alternating clicks between the TabLayout and the VerticalTextView(s) and (2) successively clicks to the same VerticalTextView was working, but (3) clicks to different VerticalTextView(s) then clicking TabLayout afterwards will cause multiple levels of fragments-associated-with-the-VerticalTextView(s) to remain in the child fragment manager's stack.

NEXT: there's an error when trying to come back from MealQueueViewerActivity (probably related to the newly added FragmentContainerView [which allowed for switching between ViewPager2 and fragments-associated-with-the-VerticalTextView(s)] of MenuItemInputFragment's layout).