refactor: Add CheckBox for queue viewer's RV item's layout
MealQueueViewerActivity displays a RecyclerView of RecordOfMeal, inside of this is another RecyclerView (this one lists MenuItem). The layout for each of the RecyclerView's MenuItem now has a CheckBox.

NEXT: validate whether all CheckBox for the meal is checked or not (if so, remove the meal). Also, add a boolean handedOff field to MenuItem and have CheckBox state be set based on this new field (currently, CheckBox do not save its checked state when scrolled off-screen and scrolled back on-screen).