feat: Remove Meal when all MenuItem have been checked off
Each menu item is tracking whether its in a handed-off state. After MenuItem are checked-off in MealQueueViewerActivity, we run a check to see if all MenuItem from this Meal have been checked (if so, remove the RecordOfMeal [the one associated with this Meal] from the list and notify the adapter about the change to its underlying data.